# scripts/sdk/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "nocodb",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/v1/db/meta/projects/{projectId}/tables": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "projectId",
          "in": "path",
          "required": true
        }
      ],
      "post": {
        "summary": "",
        "operationId": "db-table-create",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Table"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TableReq"
              }
            }
          }
        },
        "tags": [
          "DB table"
        ]
      },
      "get": {
        "summary": "",
        "operationId": "db-table-list",
        "responses": {
          "200": {
            "$ref": "#/components/responses/TableList"
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "xc-auth"
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "page"
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "pageSize"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "sort"
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "includeM2M"
          }
        ],
        "tags": [
          "DB table"
        ]
      }
    },
    "/api/v1/db/meta/projects/{projectId}/{baseId}/tables": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "projectId",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string"
          },
          "name": "baseId",
          "in": "path",
          "required": true
        }
      ],
      "post": {
        "summary": "",
        "operationId": "table-create",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Table"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TableReq"
              }
            }
          }
        },
        "tags": [
          "Base"
        ]
      }
    },
    "/api/v1/db/meta/tables/{tableId}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "tableId",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "",
        "operationId": "db-table-read",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableInfo"
                },
                "examples": {
                  "example-1": {
                    "value": {
                      "id": "string",
                      "fk_project_id": "string",
                      "fk_base_id": "string",
                      "title": "string",
                      "alias": "string",
                      "type": "string",
                      "enabled": "string",
                      "parent_id": "string",
                      "show_as": "string",
                      "tags": "string",
                      "pinned": true,
                      "deleted": true,
                      "order": 0,
                      "column": [
                        {
                          "id": "string",
                          "base_id": "string",
                          "fk_model_id": "string",
                          "title": "string",
                          "alias": "string",
                          "ui_data_type": "string",
                          "data_type": "string",
                          "numeric_precision": "string",
                          "numeric_scale": "string",
                          "character_maximum_length": "string",
                          "column_ordinal_position": "string",
                          "primary_key": "string",
                          "primary_value": "string",
                          "rqd": "string",
                          "un": "string",
                          "column_type": "string",
                          "auto_increment": "string",
                          "unique": "string",
                          "column_default": "string",
                          "column_comment": "string",
                          "character_set_name": "string",
                          "data_type_x": "string",
                          "data_type_x_precision": "string",
                          "data_type_x_scale": "string",
                          "auto_update_timestamp": true,
                          "deleted": true,
                          "visible": true,
                          "order": 0,
                          "colOptions": {
                            "id": "string",
                            "type": "string",
                            "virtual": true,
                            "fk_column_id": "string",
                            "fk_child_column_id": "string",
                            "fk_parent_column_id": "string",
                            "fk_mm_model_id": "string",
                            "fk_mm_child_column_id": "string",
                            "fk_mm_parent_column_id": "string",
                            "ur": "string",
                            "dr": "string",
                            "fk_index_name": "string",
                            "deleted": "string",
                            "order": "string"
                          }
                        }
                      ],
                      "filters": [
                        {
                          "id": "string",
                          "fk_model_id": "string",
                          "fk_column_id": "string",
                          "logical_op": "string",
                          "comparison_op": "string",
                          "value": "string",
                          "is_group": true,
                          "children": [
                            {}
                          ]
                        }
                      ],
                      "sort": [
                        {
                          "id": "string",
                          "fk_model_id": "string",
                          "fk_column_id": "string",
                          "direction": "string",
                          "order": 0
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "DB table"
        ]
      },
      "patch": {
        "summary": "",
        "operationId": "db-table-update",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "tags": [
          "DB table"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "table_name": {
                    "type": "string"
                  },
                  "project_id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "",
        "operationId": "db-table-delete",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "DB table"
        ],
        "description": ""
      }
    },
    "/api/v1/db/meta/tables/{tableId}/reorder": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "tableId",
          "in": "path",
          "required": true
        }
      ],
      "post": {
        "summary": "",
        "operationId": "db-table-reorder",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "order": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "DB table"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Table": {
        "title": "Table",
        "type": "object",
        "examples": [
          {
            "id": "string",
            "project_id": "string",
            "base_id": "string",
            "title": "string",
            "alias": "string",
            "type": "string",
            "enabled": true,
            "parent_id": "string",
            "show_as": "string",
            "tags": "string",
            "pinned": true,
            "deleted": true,
            "order": 0,
            "column": [
              {
                "id": "string",
                "base_id": "string",
                "fk_model_id": "string",
                "title": "string",
                "alias": "string",
                "ui_data_type": "string",
                "data_type": "string",
                "numeric_precision": "string",
                "numeric_scale": "string",
                "character_maximum_length": "string",
                "column_ordinal_position": "string",
                "primary_key": true,
                "primary_value": true,
                "rqd": "string",
                "un": "string",
                "column_type": "string",
                "auto_increment": true,
                "unique": true,
                "column_default": "string",
                "column_comment": "string",
                "character_set_name": "string",
                "data_type_x": "string",
                "data_type_x_precision": "string",
                "data_type_x_scale": "string",
                "auto_update_timestamp": true,
                "deleted": true,
                "visible": true,
                "order": 0,
                "colOptions": {
                  "id": "string",
                  "type": "string",
                  "virtual": true,
                  "fk_column_id": "string",
                  "fk_child_column_id": "string",
                  "fk_parent_column_id": "string",
                  "fk_mm_model_id": "string",
                  "fk_mm_child_column_id": "string",
                  "fk_mm_parent_column_id": "string",
                  "ur": "string",
                  "dr": "string",
                  "fk_index_name": "string",
                  "deleted": "string",
                  "order": "string"
                }
              }
            ],
            "columnByIds": {}
          }
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "project_id": {
            "type": "string"
          },
          "base_id": {
            "type": "string"
          },
          "table_name": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "parent_id": {
            "type": "string"
          },
          "show_as": {
            "type": "string"
          },
          "tags": {
            "type": "string"
          },
          "pinned": {
            "type": "boolean"
          },
          "deleted": {
            "type": "boolean"
          },
          "order": {
            "type": "number"
          },
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Column"
            }
          },
          "columnsById": {
            "type": "object"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "table_name",
          "title"
        ]
      },
...
      "TableInfo": {
        "title": "Table",
        "type": "object",
        "examples": [
          {
            "id": "string",
            "fk_project_id": "string",
            "fk_base_id": "string",
            "title": "string",
            "alias": "string",
            "type": "string",
            "enabled": "string",
            "parent_id": "string",
            "show_as": "string",
            "tags": "string",
            "pinned": true,
            "deleted": true,
            "order": 0,
            "column": [
              {
                "id": "string",
                "base_id": "string",
                "fk_model_id": "string",
                "title": "string",
                "alias": "string",
                "ui_data_type": "string",
                "data_type": "string",
                "numeric_precision": "string",
                "numeric_scale": "string",
                "character_maximum_length": "string",
                "column_ordinal_position": "string",
                "primary_key": "string",
                "primary_value": "string",
                "rqd": "string",
                "un": "string",
                "column_type": "string",
                "auto_increment": "string",
                "unique": "string",
                "column_default": "string",
                "column_comment": "string",
                "character_set_name": "string",
                "data_type_x": "string",
                "data_type_x_precision": "string",
                "data_type_x_scale": "string",
                "auto_update_timestamp": true,
                "deleted": true,
                "visible": true,
                "order": 0,
                "colOptions": {
                  "id": "string",
                  "type": "string",
                  "virtual": true,
                  "fk_column_id": "string",
                  "fk_child_column_id": "string",
                  "fk_parent_column_id": "string",
                  "fk_mm_model_id": "string",
                  "fk_mm_child_column_id": "string",
                  "fk_mm_parent_column_id": "string",
                  "ur": "string",
                  "dr": "string",
                  "fk_index_name": "string",
                  "deleted": "string",
                  "order": "string"
                }
              }
            ],
            "filters": [
              {
                "id": "string",
                "fk_model_id": "string",
                "fk_column_id": "string",
                "logical_op": "string",
                "comparison_op": "string",
                "value": "string",
                "is_group": true,
                "children": [
                  {}
                ]
              }
            ],
            "sort": [
              {
                "id": "string",
                "fk_model_id": "string",
                "fk_column_id": "string",
                "direction": "string",
                "order": 0
              }
            ]
          }
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "fk_project_id": {
            "type": "string"
          },
          "fk_base_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "table_name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "enabled": {
            "type": "string"
          },
          "parent_id": {
            "type": "string"
          },
          "show_as": {
            "type": "string"
          },
          "tags": {
            "type": "string"
          },
          "pinned": {
            "type": "boolean"
          },
          "deleted": {
            "type": "boolean"
          },
          "order": {
            "type": "number"
          },
          "column": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Column"
            }
          },
          "filters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Filter"
            }
          },
          "sort": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Sort"
            }
          }
        },
        "required": [
          "title",
          "table_name"
        ]
      },
      "TableReq": {
        "title": "Table",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fk_project_id": {
            "type": "string"
          },
          "fk_base_id": {
            "type": "string"
          },
          "table_name": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "enabled": {
            "type": "string"
          },
          "parent_id": {
            "type": "string"
          },
          "show_as": {
            "type": "string"
          },
          "tags": {
            "type": "string"
          },
          "pinned": {
            "type": "boolean"
          },
          "deleted": {
            "type": "boolean"
          },
          "order": {
            "type": "number"
          },
          "mm": {
            "type": "boolean"
          },
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Column"
            }
          }
        },
        "required": [
          "table_name",
          "title",
          "columns"
        ]
      },
      "TableList": {
        "description": "",
        "type": "object",
        "x-examples": {
          "example-1": {
            "bases": {
              "list": [
                {
                  "id": "string",
                  "project_id": "string",
                  "alias": "string",
                  "host": "string",
                  "port": 0,
                  "username": "string",
                  "password": "string",
                  "database": "string",
                  "url": "string",
                  "params": "string",
                  "type": "string",
                  "ssl": "string"
                }
              ],
              "pageInfo": {
                "pageSize": 0,
                "totalRows": 0,
                "sort": "string",
                "isFirstPage": true,
                "isLastPage": true
              }
            }
          }
        },
        "properties": {
          "list": {
            "type": "array",
            "uniqueItems": true,
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/Table"
            }
          },
          "pageInfo": {
            "$ref": "#/components/schemas/Paginated"
          }
        },
        "required": [
          "tables"
        ]
      },
...
      "Sort": {
        "title": "Sort",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fk_model_id": {
            "type": "string"
          },
          "fk_column_id": {
            "type": "string"
          },
          "direction": {
            "type": "string"
          },
          "order": {
            "type": "number"
          },
          "project_id": {
            "type": "string",
            "readOnly": true
          },
          "base_id": {
            "type": "string",
            "readOnly": true
          }
        }
      },
...
      "Column": {
        "title": "Column",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "base_id": {
            "type": "string"
          },
          "fk_model_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "uidt": {
            "type": "string"
          },
          "dt": {
            "type": "string"
          },
          "np": {
            "type": "string"
          },
          "ns": {
            "type": "string"
          },
          "clen": {
            "type": [
              "string",
              "integer"
            ]
          },
          "cop": {
            "type": "string"
          },
          "pk": {
            "type": "boolean"
          },
          "pv": {
            "type": "boolean"
          },
          "rqd": {
            "type": "boolean"
          },
          "column_name": {
            "type": "string"
          },
          "un": {
            "type": "boolean"
          },
          "ct": {
            "type": "string"
          },
          "ai": {
            "type": "boolean"
          },
          "unique": {
            "type": "boolean"
          },
          "cdf": {
            "type": "string"
          },
          "cc": {
            "type": "string"
          },
          "csn": {
            "type": "string"
          },
          "dtx": {
            "type": "string"
          },
          "dtxp": {
            "type": "string"
          },
          "dtxs": {
            "type": "string"
          },
          "au": {
            "type": "boolean"
          },
          "deleted": {
            "type": "boolean"
          },
          "visible": {
            "type": "boolean"
          },
          "order": {
            "type": "number"
          },
          "system": {
            "type": [
              "number",
              "boolean"
            ]
          },
          "meta": {},
          "colOptions": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/LinkToAnotherRecord"
              },
              {
                "$ref": "#/components/schemas/Formula"
              },
              {
                "$ref": "#/components/schemas/Rollup"
              },
              {
                "$ref": "#/components/schemas/Lookup"
              },
              {
                "$ref": "#/components/schemas/SelectOptions"
              },
              {
                "type": "object"
              }
            ]
          }
        },
        "required": [
          "uidt"
        ]
      },
...
    },
...
    "responses": {
...
      "TableList": {
        "description": "Example response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/TableList"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "xc-auth": {
        "name": "API Key",
        "type": "apiKey",
        "in": "header",
        "description": "JWT access token"
      },
      "xc-shared-base-id": {
        "name": "API Key",
        "type": "apiKey",
        "in": "header",
        "description": "Shared base uuid"
      }
    }
  }
}